Anim frame range import fix

+++ neverblender/nvb_anim.py
@@ -35,12 +35,12 @@ class Animation():
         new_anim.transtime = fps * self.transtime
         # new_anim.root = self.animroot
         new_anim.root_obj = noderesolver.get_obj(self.animroot, -1)
-        new_anim.frameEnd = round(fps * self.length, 0) + new_anim.frameStart
+        new_anim.frameEnd = int(round(fps * self.length, 0)) + new_anim.frameStart
         # events
         for ev_time, ev_name in self.events:
             newEvent = new_anim.eventList.add()
             newEvent.name = ev_name
-            newEvent.frame = round(fps * ev_time, 0) + new_anim.frameStart
+            newEvent.frame = int(round(fps * ev_time, 0)) + new_anim.frameStart
         # Load the animation into the objects/actions
         for node in self.nodes:
             obj = noderesolver.get_obj(node.name, node.nodeidx)
